backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic-settings==2.9.1
httpx==0.28.1
anthropic==0.52.0
apscheduler==3.11.0
cryptography==44.0.3
pyjwt==2.10.1
python-multipart==0.0.20
slack_sdk==3.34.0
aiohttp==3.11.14
